+++ src/components/form/text-input/index.js
@@ -102,6 +102,8 @@ function (_PureComponent) {
   (0, _createClass2.default)(TextInput, [{
     key: "render",
     value: function render() {
+      var _this = this;
+
       var _this$props = this.props,
           valid = _this$props.valid,
           error = _this$props.error,
@@ -114,6 +116,9 @@ function (_PureComponent) {
 
       if (tag === 'textarea') {
         return _react.default.createElement(StyledTextarea, null, _react.default.createElement(StyledInputTextarea, (0, _extends2.default)({
+          ref: function ref(input) {
+            return _this.input = input;
+          },
           valid: valid,
           error: error,
           disabled: disabled,
@@ -123,6 +128,9 @@ function (_PureComponent) {
         }, others)), _react.default.createElement(StyledGradientTextarea, null));
       } else {
         return _react.default.createElement(StyledInput, (0, _extends2.default)({
+          ref: function ref(input) {
+            return _this.input = input;
+          },
           valid: valid,
           error: error,
           disabled: disabled,